changes based on code review comments.
